Shape navigation for AR (#1)
* Sort of works?

The drawer is always open, but I want it to snap like Apple Maps

* Still can't close, an trying new things?

* Changed navigation system

The drawer is not a proper way to handle iOS navigation, so it will not be used. Instead a table view was added. Will continue to adjust and make the program look more aesthetic.

--------------------------
Maybe a small issue or not? The AR only attaches things to walls, which may be good or bad? Have yet to test it out!

* Added hints to adding objects

Now when the user taps an object button, they have a hint that tells them "tap to place"