Walls kinda working...
Everything now automatically snaps to walls. This isn't good because I want to be able to place objects in front of others (such as a hand hold in front of a volume).

Furthermore, It should continuously keep trying to find walls to find new walls when I place an object, NOT just have one.