Collision WIP
Will check whether an object can be placed and have collision. If it does then it will have a Z offset.

WORK ON: Make sure this works for all shapes, (sphere with square, etc). Have a loop to continuously check whether there's collision. For example, if the object that the user is about to place will collide, provide an offset, THEN CHECK AGAIN whether there will be collision, and keep doing this until theres a coordinate point with no collision.