Objects now automatically snap to walls (#2)
* Added SnapToWall button

The button is functional, but the logic behind is it not.

* Walls kinda working...

Everything now automatically snaps to walls. This isn't good because I want to be able to place objects in front of others (such as a hand hold in front of a volume).

Furthermore, It should continuously keep trying to find walls to find new walls when I place an object, NOT just have one.

* Collision WIP

Will check whether an object can be placed and have collision. If it does then it will have a Z offset.

WORK ON: Make sure this works for all shapes, (sphere with square, etc). Have a loop to continuously check whether there's collision. For example, if the object that the user is about to place will collide, provide an offset, THEN CHECK AGAIN whether there will be collision, and keep doing this until theres a coordinate point with no collision.

* Basic collision works!

Now objects cannot be stacked on top of one another.

However, the collision should be improved upon. Maybe instead of z axis, just find the closest location to where the user wants to place an object that has no collision? Not sure how implementation would work...